move [Global]ManifestRaw to separate header
